feat(storage): Move nlohmann/json.hpp includes to cc file
Part of the work for: #2772
