+++ google/cloud/storage/CMakeLists.txt
@@ -80,8 +80,9 @@ add_library(
     iam_policy.h
     idempotency_policy.cc
     idempotency_policy.h
-    internal/access_control_common.cc
     internal/access_control_common.h
+    internal/access_control_common_parser.cc
+    internal/access_control_common_parser.h
     internal/binary_data_as_debug_string.cc
     internal/binary_data_as_debug_string.h
     internal/bucket_acl_requests.cc
@@ -89,6 +90,7 @@ add_library(
     internal/bucket_requests.cc
     internal/bucket_requests.h
     internal/common_metadata.h
+    internal/common_metadata_parser.h
     internal/complex_option.h
     internal/compute_engine_util.cc
     internal/compute_engine_util.h
@@ -121,6 +123,8 @@ add_library(
     internal/hash_validator.h
     internal/hash_validator_impl.cc
     internal/hash_validator_impl.h
+    internal/hmac_key_metadata_parser.cc
+    internal/hmac_key_metadata_parser.h
     internal/hmac_key_requests.cc
     internal/hmac_key_requests.h
     internal/http_response.cc
@@ -131,6 +135,8 @@ add_library(
     internal/logging_resumable_upload_session.h
     internal/metadata_parser.cc
     internal/metadata_parser.h
+    internal/notification_metadata_parser.cc
+    internal/notification_metadata_parser.h
     internal/notification_requests.cc
     internal/notification_requests.h
     internal/object_acl_requests.cc
@@ -385,6 +391,7 @@ if (BUILD_TESTING)
         hashing_options_test.cc
         hmac_key_metadata_test.cc
         idempotency_policy_test.cc
+        internal/access_control_common_parser_test.cc
         internal/access_control_common_test.cc
         internal/binary_data_as_debug_string_test.cc
         internal/bucket_acl_requests_test.cc
