+++ google/cloud/storage/hmac_key_metadata_test.cc
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include "google/cloud/storage/hmac_key_metadata.h"
+#include "google/cloud/storage/internal/hmac_key_metadata_parser.h"
 #include "google/cloud/storage/internal/hmac_key_requests.h"
 #include "google/cloud/internal/format_time_point.h"
 #include <gmock/gmock.h>
