+++ google/cloud/storage/iam_policy.h
@@ -17,7 +17,9 @@
 
 #include "google/cloud/storage/version.h"
 #include "google/cloud/status_or.h"
-#include <nlohmann/json.hpp>
+#include <memory>
+#include <string>
+#include <vector>
 
 namespace google {
 namespace cloud {
