+++ google/cloud/storage/internal/hmac_key_metadata_parser.cc
@@ -1,4 +1,4 @@
-// Copyright 2018 Google LLC
+// Copyright 2020 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,37 +12,42 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "google/cloud/storage/internal/access_control_common.h"
-#include <nlohmann/json.hpp>
+#include "google/cloud/storage/internal/hmac_key_metadata_parser.h"
+#include "google/cloud/internal/parse_rfc3339.h"
 
 namespace google {
 namespace cloud {
 namespace storage {
 inline namespace STORAGE_CLIENT_NS {
 namespace internal {
-Status AccessControlCommon::ParseFromJson(AccessControlCommon& result,
-                                          nlohmann::json const& json) {
+StatusOr<HmacKeyMetadata> HmacKeyMetadataParser::FromJson(
+    nlohmann::json const& json) {
   if (!json.is_object()) {
     return Status(StatusCode::kInvalidArgument, __func__);
   }
-  result.bucket_ = json.value("bucket", "");
-  result.domain_ = json.value("domain", "");
-  result.email_ = json.value("email", "");
-  result.entity_ = json.value("entity", "");
-  result.entity_id_ = json.value("entityId", "");
+  HmacKeyMetadata result{};
+  result.access_id_ = json.value("accessId", "");
   result.etag_ = json.value("etag", "");
   result.id_ = json.value("id", "");
   result.kind_ = json.value("kind", "");
-  result.role_ = json.value("role", "");
-  result.self_link_ = json.value("selfLink", "");
-  if (json.count("projectTeam") != 0) {
-    auto tmp = json["projectTeam"];
-    ProjectTeam p;
-    p.project_number = tmp.value("projectNumber", "");
-    p.team = tmp.value("team", "");
-    result.project_team_ = std::move(p);
+  result.project_id_ = json.value("projectId", "");
+  result.service_account_email_ = json.value("serviceAccountEmail", "");
+  result.state_ = json.value("state", "");
+  if (json.count("timeCreated") != 0) {
+    result.time_created_ =
+        google::cloud::internal::ParseRfc3339(json.value("timeCreated", ""));
   }
-  return Status();
+  if (json.count("updated") != 0) {
+    result.updated_ =
+        google::cloud::internal::ParseRfc3339(json.value("updated", ""));
+  }
+  return result;
+}
+
+StatusOr<HmacKeyMetadata> HmacKeyMetadataParser::FromString(
+    std::string const& payload) {
+  auto json = nlohmann::json::parse(payload, nullptr, false);
+  return FromJson(json);
 }
 
 }  // namespace internal
