+++ google/cloud/storage/internal/hmac_key_requests.h
@@ -29,11 +29,6 @@ namespace storage {
 inline namespace STORAGE_CLIENT_NS {
 namespace internal {
 
-struct HmacKeyMetadataParser {
-  static StatusOr<HmacKeyMetadata> FromJson(nlohmann::json const& json);
-  static StatusOr<HmacKeyMetadata> FromString(std::string const& payload);
-};
-
 template <typename Derived, typename... Options>
 class GenericHmacKeyRequest : public GenericRequest<Derived, Options...> {
  public:
