+++ google/cloud/storage/internal/hmac_key_requests_test.cc
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include "google/cloud/storage/internal/hmac_key_requests.h"
+#include "google/cloud/storage/internal/hmac_key_metadata_parser.h"
 #include "google/cloud/testing_util/assert_ok.h"
 #include <gmock/gmock.h>
