+++ google/cloud/storage/internal/notification_requests.cc
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 #include "google/cloud/storage/internal/notification_requests.h"
+#include "google/cloud/storage/internal/notification_metadata_parser.h"
 #include <iostream>
 
 namespace google {
@@ -20,43 +21,6 @@ namespace cloud {
 namespace storage {
 inline namespace STORAGE_CLIENT_NS {
 namespace internal {
-StatusOr<NotificationMetadata> NotificationMetadataParser::FromJson(
-    nlohmann::json const& json) {
-  if (!json.is_object()) {
-    return Status(StatusCode::kInvalidArgument, __func__);
-  }
-  NotificationMetadata result{};
-
-  if (json.count("custom_attributes") != 0) {
-    for (auto const& kv : json["custom_attributes"].items()) {
-      result.custom_attributes_.emplace(kv.key(),
-                                        kv.value().get<std::string>());
-    }
-  }
-  result.etag_ = json.value("etag", "");
-
-  if (json.count("event_types") != 0) {
-    for (auto const& kv : json["event_types"].items()) {
-      result.event_types_.emplace_back(kv.value().get<std::string>());
-    }
-  }
-
-  result.id_ = json.value("id", "");
-  result.kind_ = json.value("kind", "");
-  result.object_name_prefix_ = json.value("object_name_prefix", "");
-  result.payload_format_ = json.value("payload_format", "");
-  result.self_link_ = json.value("selfLink", "");
-  result.topic_ = json.value("topic", "");
-
-  return result;
-}
-
-StatusOr<NotificationMetadata> NotificationMetadataParser::FromString(
-    std::string const& payload) {
-  auto json = nlohmann::json::parse(payload, nullptr, false);
-  return FromJson(json);
-}
-
 std::ostream& operator<<(std::ostream& os, ListNotificationsRequest const& r) {
   os << "ListNotificationsRequest={bucket_name=" << r.bucket_name();
   r.DumpOptions(os, ", ");
