+++ google/cloud/storage/internal/notification_requests.h
@@ -20,7 +20,6 @@
 #include "google/cloud/storage/notification_metadata.h"
 #include "google/cloud/storage/version.h"
 #include "google/cloud/storage/well_known_parameters.h"
-#include <nlohmann/json.hpp>
 #include <iosfwd>
 
 namespace google {
@@ -28,11 +27,6 @@ namespace cloud {
 namespace storage {
 inline namespace STORAGE_CLIENT_NS {
 namespace internal {
-struct NotificationMetadataParser {
-  static StatusOr<NotificationMetadata> FromJson(nlohmann::json const& json);
-  static StatusOr<NotificationMetadata> FromString(std::string const& payload);
-};
-
 /// Represents a request to call the `BucketAccessControls: list` API.
 class ListNotificationsRequest
     : public GenericRequest<ListNotificationsRequest, UserProject> {
