+++ google/cloud/storage/notification_metadata.cc
@@ -15,6 +15,7 @@
 #include "google/cloud/storage/notification_metadata.h"
 #include "google/cloud/storage/internal/notification_requests.h"
 #include "google/cloud/internal/absl_str_join_quiet.h"
+#include <nlohmann/json.hpp>
 
 namespace google {
 namespace cloud {
