+++ google/cloud/storage/storage_client.bzl
@@ -27,10 +27,12 @@ storage_client_hdrs = [
     "iam_policy.h",
     "idempotency_policy.h",
     "internal/access_control_common.h",
+    "internal/access_control_common_parser.h",
     "internal/binary_data_as_debug_string.h",
     "internal/bucket_acl_requests.h",
     "internal/bucket_requests.h",
     "internal/common_metadata.h",
+    "internal/common_metadata_parser.h",
     "internal/complex_option.h",
     "internal/compute_engine_util.h",
     "internal/const_buffer.h",
@@ -49,11 +51,13 @@ storage_client_hdrs = [
     "internal/generic_request.h",
     "internal/hash_validator.h",
     "internal/hash_validator_impl.h",
+    "internal/hmac_key_metadata_parser.h",
     "internal/hmac_key_requests.h",
     "internal/http_response.h",
     "internal/logging_client.h",
     "internal/logging_resumable_upload_session.h",
     "internal/metadata_parser.h",
+    "internal/notification_metadata_parser.h",
     "internal/notification_requests.h",
     "internal/object_acl_requests.h",
     "internal/object_read_source.h",
@@ -119,7 +123,7 @@ storage_client_srcs = [
     "hmac_key_metadata.cc",
     "iam_policy.cc",
     "idempotency_policy.cc",
-    "internal/access_control_common.cc",
+    "internal/access_control_common_parser.cc",
     "internal/binary_data_as_debug_string.cc",
     "internal/bucket_acl_requests.cc",
     "internal/bucket_requests.cc",
@@ -137,11 +141,13 @@ storage_client_srcs = [
     "internal/empty_response.cc",
     "internal/hash_validator.cc",
     "internal/hash_validator_impl.cc",
+    "internal/hmac_key_metadata_parser.cc",
     "internal/hmac_key_requests.cc",
     "internal/http_response.cc",
     "internal/logging_client.cc",
     "internal/logging_resumable_upload_session.cc",
     "internal/metadata_parser.cc",
+    "internal/notification_metadata_parser.cc",
     "internal/notification_requests.cc",
     "internal/object_acl_requests.cc",
     "internal/object_requests.cc",
