+++ google/cloud/storage/storage_client_unit_tests.bzl
@@ -34,6 +34,7 @@ storage_client_unit_tests = [
     "hashing_options_test.cc",
     "hmac_key_metadata_test.cc",
     "idempotency_policy_test.cc",
+    "internal/access_control_common_parser_test.cc",
     "internal/access_control_common_test.cc",
     "internal/binary_data_as_debug_string_test.cc",
     "internal/bucket_acl_requests_test.cc",
